Switch gsl::span from GSL to tcbrindle’s implementation

+++ rpcs3/Emu/RSX/GL/GLGSRender.cpp
@@ -1372,7 +1372,7 @@ void GLGSRender::load_program_env()
 		auto mapping = m_fragment_constants_buffer->alloc_from_heap(fragment_constants_size, m_uniform_buffer_offset_align);
 		auto buf = static_cast<u8*>(mapping.first);
 
-		m_prog_buffer.fill_fragment_constants_buffer({ reinterpret_cast<float*>(buf), gsl::narrow<int>(fragment_constants_size) },
+		m_prog_buffer.fill_fragment_constants_buffer({ reinterpret_cast<float*>(buf), fragment_constants_size },
 			current_fragment_program, gl::get_driver_caps().vendor_NVIDIA);
 
 		m_fragment_constants_buffer->bind_range(GL_FRAGMENT_CONSTANT_BUFFERS_BIND_SLOT, mapping.second, fragment_constants_size);
